moved test files to accomodate postgress
