+++ usaspending_api/search/filters/elasticsearch/naics.py
@@ -1,6 +1,6 @@
 from usaspending_api.common.exceptions import InvalidParameterException
-from usaspending_api.search.elasticsearch.filters.filter import _Filter, _QueryType
-from usaspending_api.search.elasticsearch.filters.HierarchicalFilter import HierarchicalFilter, Node
+from usaspending_api.search.filters.elasticsearch.filter import _Filter, _QueryType
+from usaspending_api.search.filters.elasticsearch.HierarchicalFilter import HierarchicalFilter, Node
 from elasticsearch_dsl import Q as ES_Q
