+++ usaspending_api/search/tests/integration/hierarchical_filters/es_search_test_helpers.py
@@ -1,6 +1,6 @@
 import json
 
-from usaspending_api.search.elasticsearch.filters.tas import TasCodes, TreasuryAccounts
+from usaspending_api.search.filters.elasticsearch.tas import TasCodes, TreasuryAccounts
 from usaspending_api.common.experimental_api_flags import EXPERIMENTAL_API_HEADER, ELASTICSEARCH_HEADER_VALUE
 from usaspending_api.search.tests.data.utilities import setup_elasticsearch_test
